refactor: split main.go into focused files by responsibility
- main.go: entry point, MisuseError, logger
- cmd_root.go: rootCmd, aiCmd, global flag vars, init()
- cmd_file.go: fileCmd + aiFileCmd
- cmd_notification.go: notificationCmd + aiNotificationCmd
- cmd_support_packet.go: supportPacketCmd + aiSupportPacketCmd
- cmd_version.go: versionCmd
- flags.go: flag registration helpers and validation
- logs.go: loadFileLogs, processLogs, runAIAnalysis
